Move X11 specific code from Display to DisplayServer
A step forward for a display server agnostic sddm.
With Wayland support coming soon we want display server specific code
not scattered all over the place.
