+++ src/daemon/Display.h
@@ -47,9 +47,6 @@ namespace SDDM {
 
         const QString &name() const;
 
-        const QString &cookie() const;
-        void addCookie(const QString &file);
-
         Seat *seat() const;
 
     public slots:
@@ -74,9 +71,6 @@ namespace SDDM {
         int m_displayId { 0 };
         int m_terminalId { 7 };
 
-        QString m_display { ":0" };
-        QString m_cookie { "" };
-        QString m_authPath { "" };
         QString m_passPhrase;
         QString m_sessionName;
